[Gold IV] Title: 수 묶기, Time: 120 ms, Memory: 14232 KB -BaekjoonHub

+++ 백준/Gold/1744. 수 묶기/수 묶기.java
@@ -11,7 +11,7 @@ public class Main {
 
         int N = Integer.parseInt(br.readLine());
         int[] arr = new int[N];
-        long answer = 0;
+        int answer = 0;
 
         for (int i = 0; i < N; i++) {
             st = new StringTokenizer(br.readLine());
